refactor: modularize sorting logic for reusability
